refactor(useValueRef): import useValueRef from maskbook-shared directly
