+++ packages/maskbook/src/components/InjectedComponents/PostDialogHint.tsx
@@ -4,7 +4,7 @@ import { makeStyles } from '@material-ui/core/styles'
 import { useI18N } from '../../utils'
 import { useMyIdentities } from '../DataSource/useActivatedUI'
 import type { BannerProps } from '../Welcomes/Banner'
-import { useValueRef } from '../../utils/hooks/useValueRef'
+import { useValueRef } from '@dimensiondev/maskbook-shared'
 import { currentSetupGuideStatus } from '../../settings/settings'
 import { activatedSocialNetworkUI } from '../../social-network'
 import { isMobileFacebook } from '../../social-network-adaptor/facebook.com/utils/isMobile'
