+++ packages/maskbook/src/components/InjectedComponents/SetupGuide.tsx
@@ -22,7 +22,8 @@ import ArrowBackIosOutlinedIcon from '@material-ui/icons/ArrowBackIosOutlined'
 import stringify from 'json-stable-stringify'
 import ActionButton, { ActionButtonPromise } from '../../extension/options-page/DashboardComponents/ActionButton'
 import { noop } from 'lodash-es'
-import { useI18N, MaskMessage, useValueRef, useMatchXS, extendsTheme } from '../../utils'
+import { useValueRef } from '@dimensiondev/maskbook-shared'
+import { useI18N, MaskMessage, useMatchXS, extendsTheme } from '../../utils'
 import { activatedSocialNetworkUI } from '../../social-network'
 import { currentSetupGuideStatus } from '../../settings/settings'
 import type { SetupGuideCrossContextStatus } from '../../settings/types'
