+++ packages/maskbook/src/components/shared-settings/useSettingsUI.tsx
@@ -1,21 +1,20 @@
 import type { ValueRef } from '@dimensiondev/holoflows-kit'
-import { useValueRef } from '../../utils/hooks/useValueRef'
-import { useMatchXS } from '../../utils/hooks/useMatchXS'
-import { texts } from '../../settings/createSettings'
+import { useValueRef } from '@dimensiondev/maskbook-shared'
 import {
     ListItem,
-    ListItemText,
-    ListItemSecondaryAction,
-    Switch,
-    Select,
-    MenuItem,
-    makeStyles,
     ListItemIcon,
+    ListItemSecondaryAction,
+    ListItemText,
+    makeStyles,
+    MenuItem,
+    Select,
     SelectProps,
+    Switch,
 } from '@material-ui/core'
 import ArrowForwardIosIcon from '@material-ui/icons/ArrowForwardIos'
+import { texts } from '../../settings/createSettings'
+import { getEnumAsArray, useMatchXS } from '../../utils'
 import { useStylesExtends } from '../custom-ui-helper'
-import { getEnumAsArray } from '../../utils/enum'
 
 const useStyles = makeStyles((theme) => ({
     container: { listStyleType: 'none', width: '100%' },
