+++ packages/maskbook/src/plugins/Collectible/hooks/useCollectibleState.ts
@@ -2,7 +2,7 @@ import { useRef, useState } from 'react'
 import { useUpdateEffect } from 'react-use'
 import { createContainer } from 'unstated-next'
 import { OrderSide } from 'opensea-js/lib/types'
-import { useValueRef } from '../../../utils/hooks/useValueRef'
+import { useValueRef } from '@dimensiondev/maskbook-shared'
 import { currentCollectibleProviderSettings } from '../settings'
 import { CollectibleTab, CollectibleToken } from '../types'
 import { useAsset } from './useAsset'
