+++ packages/maskbook/src/plugins/Polls/UI/Polls.tsx
@@ -2,7 +2,8 @@ import { makeStyles, Card, Typography, CircularProgress, List, ListItem } from '
 import isValid from 'date-fns/isValid'
 import formatDistance from 'date-fns/formatDistance'
 import { zhTW, enUS, ja } from 'date-fns/locale'
-import { useI18N, useValueRef } from '../../../utils'
+import { useValueRef } from '@dimensiondev/maskbook-shared'
+import { useI18N } from '../../../utils'
 import { languageSettings } from '../../../settings/settings'
 import type { PollGunDB } from '../Services'
 import { PollStatus } from '../types'
