+++ packages/maskbook/src/plugins/Trader/UI/trader/SettingsDialog.tsx
@@ -12,7 +12,8 @@ import {
     Button,
     DialogContent,
 } from '@material-ui/core'
-import { getEnumAsArray, useI18N, useRemoteControlledDialog, useValueRef } from '../../../../utils'
+import { useValueRef } from '@dimensiondev/maskbook-shared'
+import { getEnumAsArray, useI18N, useRemoteControlledDialog } from '../../../../utils'
 import { ZrxTradePool, TradeProvider } from '../../types'
 import { SelectPoolPanel } from './SelectPoolPanel'
 import { SlippageSlider } from './SlippageSlider'
