+++ packages/maskbook/src/plugins/Trader/UI/trending/TrendingViewDeck.tsx
@@ -4,8 +4,8 @@ import MonetizationOnOutlinedIcon from '@material-ui/icons/MonetizationOnOutline
 import ArrowDropDownIcon from '@material-ui/icons/ArrowDropDown'
 import stringify from 'json-stable-stringify'
 import { first, last } from 'lodash-es'
-import { FormattedCurrency } from '@dimensiondev/maskbook-shared'
-import { useRemoteControlledDialog, useI18N, useValueRef, Flags } from '../../../../utils'
+import { FormattedCurrency, useValueRef } from '@dimensiondev/maskbook-shared'
+import { useRemoteControlledDialog, useI18N, Flags } from '../../../../utils'
 import { Coin, Currency, DataProvider, Stat, TradeProvider, Trending } from '../../types'
 import { PriceChanged } from './PriceChanged'
 import { Linking } from './Linking'
