+++ packages/maskbook/src/plugins/Trader/trader/0x/useSlippageTolerance.ts
@@ -1,5 +1,5 @@
 import { clamp } from 'lodash-es'
-import { useValueRef } from '../../../../utils/hooks/useValueRef'
+import { useValueRef } from '@dimensiondev/maskbook-shared'
 import { SLIPPAGE_TOLERANCE_MAX, SLIPPAGE_TOLERANCE_MIN } from '../../constants'
 import { currentSlippageTolerance } from '../../settings'
