+++ packages/maskbook/src/plugins/Trader/trending/useCurrentCoinId.ts
@@ -1,5 +1,5 @@
 import { useMemo } from 'react'
-import { useValueRef } from '../../../utils/hooks/useValueRef'
+import { useValueRef } from '@dimensiondev/maskbook-shared'
 import { getCurrentPreferredCoinIdSettings } from '../settings'
 import type { DataProvider } from '../types'
