+++ packages/maskbook/src/plugins/Wallet/hooks/useERC1155Tokens.ts
@@ -1,10 +1,10 @@
 import { ValueRef } from '@dimensiondev/holoflows-kit'
-import { WalletMessages, WalletRPC } from '../messages'
-import { useValueRef } from '../../../utils/hooks/useValueRef'
+import { useValueRef } from '@dimensiondev/maskbook-shared'
+import { ERC1155TokenDetailed, EthereumTokenType } from '@dimensiondev/web3-shared'
 import type { ERC1155TokenRecord } from '../database/types'
 import { ERC1155TokenArrayComparer } from '../helpers'
+import { WalletMessages, WalletRPC } from '../messages'
 import { useWallet } from './useWallet'
-import { ERC1155TokenDetailed, EthereumTokenType } from '@dimensiondev/web3-shared'
 
 //#region cache service query result
 const erc1155TokensRef = new ValueRef<ERC1155TokenRecord[]>([], ERC1155TokenArrayComparer)
