+++ packages/maskbook/src/plugins/utils/createCompositionDialog.tsx
@@ -1,5 +1,5 @@
 import { ValueRef } from '@dimensiondev/holoflows-kit'
-import { useValueRef } from '../../utils/hooks/useValueRef'
+import { useValueRef } from '@dimensiondev/maskbook-shared'
 import type { PluginConfig } from '../types'
 
 export function createCompositionDialog(
