+++ packages/maskbook/src/utils/theme.ts
@@ -1,17 +1,16 @@
-import { unstable_createMuiStrictModeTheme, useMediaQuery } from '@material-ui/core'
-import { makeStyles } from '@material-ui/core/styles'
-import { orange, green, red, blue, grey } from '@material-ui/core/colors'
-import type { Theme, ThemeOptions } from '@material-ui/core/styles/createTheme'
-import { merge, cloneDeep } from 'lodash-es'
-import { appearanceSettings, languageSettings } from '../settings/settings'
-import { useValueRef } from './hooks/useValueRef'
-import { useMemo, useRef } from 'react'
-import { zhTW, koKR, jaJP } from '@material-ui/core/locale/index'
-import { safeUnreachable } from '@dimensiondev/maskbook-shared'
-import { or } from '../components/custom-ui-helper'
-import { activatedSocialNetworkUI } from '../social-network'
 import { ValueRef } from '@dimensiondev/holoflows-kit'
+import { safeUnreachable, useValueRef } from '@dimensiondev/maskbook-shared'
 import { Appearance, Language } from '@dimensiondev/maskbook-theme'
+import { unstable_createMuiStrictModeTheme, useMediaQuery } from '@material-ui/core'
+import { blue, green, grey, orange, red } from '@material-ui/core/colors'
+import { jaJP, koKR, zhTW } from '@material-ui/core/locale/index'
+import { makeStyles } from '@material-ui/core/styles'
+import type { Theme, ThemeOptions } from '@material-ui/core/styles/createTheme'
+import { cloneDeep, merge } from 'lodash-es'
+import { useMemo, useRef } from 'react'
+import { or } from '../components/custom-ui-helper'
+import { appearanceSettings, languageSettings } from '../settings/settings'
+import { activatedSocialNetworkUI } from '../social-network'
 
 function getFontFamily(monospace?: boolean) {
     // We want to look native.
